Rename README.md to Code
This is a python code where you can mae your own Python Virtual Assiatant.

This code is easyly expandable. This can be done by adding an elif block.

This code requires a few modules from which a few need to be installed into python. The istallation process is quite simple. 
Type the folowing code in the terminal of your python one by one
1. pip install pyttsx3
2. pip install wikipedia
3. pip install pyjokes
4. pip install randfacts
5. pip install SpeechRecognition

After the installation of the Python modules, just copy the code and run it. Any recomendations and suggestions are welcome